CB-8420 'cordova plugin add' should retrieve version, url or local folder from config.xml if none is specified
